requirements.txt
fastapi>=0.68.0, <0.69.0
uvicorn>=0.15.0, <0.16.0
sqlalchemy>=1.4.22, <1.5.0
passlib>=1.7.4, <1.8.0
python-dotenv>=0.19.0, <0.20.0
python-jose>=3.3.0,<3.4.0
rsa>=4.7.2,<4.8.0
cryptography>=3.4.7<3.5.0
python-multipart>=0.0.5,<0.1.0